moved scrolling logic from ConversationBody to the Conversation component
