+++ client/src/ui/messaging/ConversationBody.tsx
@@ -1,8 +1,8 @@
 "use client";
 import { MessageCard, MessageCardProps } from "./MessageCard";
-import { useRef, useEffect, memo } from "react";
+import { useRef, memo } from "react";
 import { NewMessagesNotification } from "./NewMessagesNotification";
-
+import { MutableRefObject } from "react";
 /**
  * Handles rendering messages
  * @param messages: Messages sent from users in a particular conversation
@@ -11,30 +11,21 @@ export const ConversationBody = memo(function ConversationBody({
   messages,
   numNewMessages,
   resetNumNewMessages,
+  scrollDown,
+  bottomRef,
 }: {
   messages: MessageCardProps[];
   numNewMessages: number;
   resetNumNewMessages: () => void;
+  scrollDown: () => void;
+  bottomRef: MutableRefObject<HTMLDivElement | null>;
 }) {
   const prevDate = useRef<Date | null>(null); // used for rendering date divider
-  const bottomRef = useRef<null | HTMLDivElement>(null); // used for scrolling down the page
-
-  // Scroll to the bottom of the element
-  function scrollDown() {
-    if (bottomRef.current) {
-      bottomRef.current.scrollIntoView({ behavior: "smooth" });
-    }
-  }
-
-  // scrolls down to the latest message on page mount
-  useEffect(() => {
-    scrollDown();
-  }, []); // Empty dependency array ensures this runs only once on mount
 
   return (
     <div
       style={{
-        overflowY: "auto", // allows scrolling on the messages
+        overflowY: "auto", // allows vertical scrolling on the messages
       }}
     >
       <NewMessagesNotification
